IGNITE-16295 remove old classes
